Refactor tutorial into a separate file
