+++ frontend/src/components/CookRecipe.js
@@ -57,30 +57,20 @@
 // THE SOFTWARE.
 
 import Button from "react-bootstrap/Button";
-import Carousel from "react-bootstrap/Carousel";
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import Tab from "react-bootstrap/Tab";
 import Tabs from "react-bootstrap/Tabs";
 import "./CookRecipe.css";
-import navigateNext from "../icons/navigate_next.svg";
 import navigatePrevious from "../icons/navigate_previous.svg";
-import speakerOn from "../icons/speaker-on.svg";
-import speakerOff from "../icons/speaker-off.svg";
+import Tutorial from "./Tutorial";
 
 class CookRecipe extends Component {
   constructor(properties) {
     super(properties);
     this.state = {
       recipe: JSON.parse(localStorage.getItem("recipe")),
-      isLastStep: false,
-      isSpeakerOff: true,
     };
-    this.switchSpeaker = this.switchSpeaker.bind(this);
-  }
-
-  componentDidMount() {
-    this.noteIfLastStep();
   }
 
   render() {
@@ -115,97 +105,12 @@ class CookRecipe extends Component {
           </Tab>
           <Tab eventKey="tutorial" title="Tutorial">
             <div className="tab-content">
-              <div className="centered-div">
-                <Button
-                  variant="primary"
-                  onClick={this.switchSpeaker}
-                >
-                  <img src={this.getSpeakerIcon()} alt="switch speaker" />
-                  <div></div>
-                  {this.getSpeakerMessage()}
-                </Button>
-              </div>
-
-              <Carousel
-                interval={null} // to disable auto play of the carousel
-                onSelect={this.setSelectedStepAndMaybeRead}
-                defaultActiveIndex={this.getSelectedStep}
-                nextIcon={
-                  <img
-                    src={navigateNext}
-                    alt="next step"
-                    className="carousel-control"
-                  />
-                }
-                prevIcon={
-                  <img
-                    src={navigatePrevious}
-                    alt="previous step"
-                    className="carousel-control"
-                  />
-                }
-              >
-                {recipe.instructions.map((step, i) => (
-                  <Carousel.Item key={i} className="carousel-step">
-                    {step}
-                  </Carousel.Item>
-                ))}
-              </Carousel>
-              {this.displayFinishedIfLastStep()}
+              <Tutorial recipe={this.state.recipe}/>
             </div>
           </Tab>
         </Tabs>
       </div>
     );
-  }
-
-  setSelectedStepAndMaybeRead = (selectedIndex, e) => {
-    localStorage.setItem("tutorial-step", selectedIndex);
-    this.noteIfLastStep();
-    if (this.state.isSpeakerOff) {
-      return;
-    }
-    // to do : read the step
-  };
-
-  getSelectedStep() {
-    const step = JSON.parse(localStorage.getItem("tutorial-step"));
-    return step ? step : 0;
-  }
-
-  noteIfLastStep() {
-    const step = this.getSelectedStep();
-    const isLastStep = this.state.recipe.instructions.length === step + 1;
-    this.setState({ isLastStep: isLastStep });
-  }
-
-  displayFinishedIfLastStep() {
-    if (this.state.isLastStep) {
-      return (
-        <div className="centered-div">
-          <Button onClick={this.finishCooking}>All done!</Button>
-        </div>
-      );
-    }
-  }
-
-  switchSpeaker() {
-    const previousStateIsSpeakerOff = this.state.isSpeakerOff;
-    this.setState({ isSpeakerOff: !previousStateIsSpeakerOff });
-  }
-
-  finishCooking() {
-    // to do
-  }
-
-  getSpeakerIcon() {
-    return this.state.isSpeakerOff ? speakerOn : speakerOff;
-  }
-
-  getSpeakerMessage() {
-    return this.state.isSpeakerOff
-      ? "I want to hear the instructions"
-      : "Don't read the instructions";
   }
 }
 export default CookRecipe;
