+++ frontend/src/components/Tutorial.js
@@ -1,0 +1,126 @@
+// Copyright 2020 Google LLC
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+
+//     https://www.apache.org/licenses/LICENSE-2.0
+
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import Button from "react-bootstrap/Button";
+import Carousel from "react-bootstrap/Carousel";
+import React, { Component } from "react";
+import "./CookRecipe.css";
+import navigateNext from "../icons/navigate_next.svg";
+import navigatePrevious from "../icons/navigate_previous.svg";
+import speakerOn from "../icons/speaker-on.svg";
+import speakerOff from "../icons/speaker-off.svg";
+
+class Tutorial extends Component {
+  constructor(properties) {
+    super(properties);
+    this.state = {
+      isLastStep: false,
+      isSpeakerOff: true,
+    };
+    this.switchSpeaker = this.switchSpeaker.bind(this);
+  }
+
+  componentDidMount() {
+    this.noteIfLastStep();
+  }
+
+  render() {
+    return (
+      <div>
+        <div className="centered-div">
+          <Button variant="primary" onClick={this.switchSpeaker}>
+            <img src={this.getSpeakerIcon()} alt="switch speaker" />
+            {this.getSpeakerMessage()}
+          </Button>
+        </div>
+
+        <Carousel
+          interval={null} // to disable auto play of the carousel
+          onSelect={this.setSelectedStepAndMaybeRead}
+          defaultActiveIndex={this.getSelectedStep}
+          nextIcon={
+            <img
+              src={navigateNext}
+              alt="next step"
+              className="carousel-control"
+            />
+          }
+          prevIcon={
+            <img
+              src={navigatePrevious}
+              alt="previous step"
+              className="carousel-control"
+            />
+          }
+        >
+          {this.props.recipe.instructions.map((step, i) => (
+            <Carousel.Item key={i} className="carousel-step">
+              {step}
+            </Carousel.Item>
+          ))}
+        </Carousel>
+        {this.displayFinishedIfLastStep()}
+      </div>
+    );
+  }
+  setSelectedStepAndMaybeRead = (selectedIndex, e) => {
+    localStorage.setItem("tutorial-step", selectedIndex);
+    this.noteIfLastStep();
+    if (this.state.isSpeakerOff) {
+      return;
+    }
+    // to do : read the step
+  };
+
+  getSelectedStep() {
+    const step = JSON.parse(localStorage.getItem("tutorial-step"));
+    return step ? step : 0;
+  }
+
+  noteIfLastStep() {
+    const step = this.getSelectedStep();
+    const isLastStep = this.props.recipe.instructions.length === step + 1;
+    this.setState({ isLastStep: isLastStep });
+  }
+
+  displayFinishedIfLastStep() {
+    if (this.state.isLastStep) {
+      return (
+        <div className="centered-div">
+          <Button onClick={this.finishCooking}>All done!</Button>
+        </div>
+      );
+    }
+  }
+
+  switchSpeaker() {
+    const previousStateIsSpeakerOff = this.state.isSpeakerOff;
+    this.setState({ isSpeakerOff: !previousStateIsSpeakerOff });
+  }
+
+  finishCooking() {
+    // to do
+  }
+
+  getSpeakerIcon() {
+    return this.state.isSpeakerOff ? speakerOn : speakerOff;
+  }
+
+  getSpeakerMessage() {
+    return this.state.isSpeakerOff
+      ? "I want to hear the instructions"
+      : "Don't read the instructions";
+  }
+}
+export default Tutorial;
